% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basketball}
\alias{basketball}
\title{Basketball results}
\format{An object of class \code{data.frame} with 4083 rows and 22 columns.}
\usage{
basketball
}
\description{
Basketball results
}
\keyword{datasets}
