% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fbs}
\alias{fbs}
\title{Subset of NCAA in FBS}
\format{An object of class \code{data.frame} with 1343 rows and 43 columns.}
\usage{
fbs
}
\description{
Subset of NCAA in FBS
}
\keyword{datasets}
