% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{football}
\alias{football}
\title{Football results}
\format{An object of class \code{data.frame} with 1471 rows and 22 columns.}
\usage{
football
}
\description{
Football results
}
\keyword{datasets}
