% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sports-ref.R
\name{get_fball_year}
\alias{get_fball_year}
\alias{get_bball_year}
\title{Scrape football and basketball data}
\usage{
get_fball_year(year)

get_bball_year(year)
}
\arguments{
\item{year}{Year to scrape}
}
\description{
Scrape football and basketball data
}
