% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_graduates}
\alias{hs_graduates}
\title{High School graduates by state}
\format{An object of class \code{data.frame} with 630 rows and 3 columns.}
\usage{
hs_graduates
}
\description{
High School graduates by state
}
\keyword{datasets}
