% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incomes}
\alias{incomes}
\title{Income by state}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 572 rows and 6 columns.}
\usage{
incomes
}
\description{
Income by state
}
\keyword{datasets}
