% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{institutions}
\alias{institutions}
\title{IPEDS and EADA data}
\format{An object of class \code{data.frame} with 1648 rows and 30 columns.}
\usage{
institutions
}
\description{
IPEDS and EADA data
}
\keyword{datasets}
