% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb.R
\name{load_uscb}
\alias{load_uscb}
\title{Load data from the USCB}
\usage{
load_uscb(file, ...)
}
\arguments{
\item{file}{path to the CSV file to load}

\item{...}{currently ignored}
}
\description{
Load data from the USCB
}
