% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{power5}
\alias{power5}
\title{Subset of FBS in Power Five conferences}
\format{An object of class \code{data.frame} with 715 rows and 45 columns.}
\usage{
power5
}
\description{
Subset of FBS in Power Five conferences
}
\keyword{datasets}
