% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schools}
\alias{schools}
\title{Lookup table for school names}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 351 rows and 2 columns.}
\usage{
schools
}
\description{
Lookup table for school names
}
\keyword{datasets}
