% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sports}
\alias{sports}
\title{Basketball and Football results}
\format{An object of class \code{data.frame} with 4083 rows and 50 columns.}
\usage{
sports
}
\description{
Basketball and Football results
}
\keyword{datasets}
