% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{Income and HS graduation by state}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 568 rows and 9 columns.}
\usage{
states
}
\description{
Income and HS graduation by state
}
\keyword{datasets}
