% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{donations}
\alias{donations}
\title{VSE Donations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6044 rows and 9 columns.}
\usage{
donations
}
\description{
VSE Donations
}
\keyword{datasets}
